cpi-viewer-container/requirements.txt
python-dotenv
requests
Flask
flask_cors
saspy
gunicorn